+++ front/src/common/components/Table.module.scss
@@ -1,17 +1,29 @@
-.table {
+@import "src/scss/Constants.scss";
+
+.Table {
 }
 
-.thead {
+.TableHead {
+  background-color: $grey-light;
 }
 
-.tbody {
+.TableBody {
 }
 
-.tr {
+.TableRow {
 }
 
-.th {
+.TableHeaderCell,
+.TableCell {
+  font-size: 14px;
+  padding: 8px 20px;
+  border-bottom: 1px solid $grey-light;
 }
 
-.td {
+.TableHeaderCell {
+  font-weight: bold;
+  text-align: left;
+}
+
+.TableCell {
 }
